+++ tests/xso/test_model.py
@@ -4350,6 +4350,430 @@ class ChildTag(XMLTestCase):
         del self.prop
 
 
+class TestChildValueList(unittest.TestCase):
+    class ChildXSO(xso.XSO):
+        TAG = ("uri:foo", "foo")
+
+        attr = xso.Attr("a", type_=xso.Integer())
+
+    class ChildValueType(xso.AbstractType):
+        @classmethod
+        def get_formatted_type(cls):
+            return TestChildValueList.ChildXSO
+
+        def coerce(self, value):
+            return TestChildValueList.ChildXSO.attr.type_.coerce(value)
+
+        def format(self, value):
+            item = TestChildValueList.ChildXSO()
+            item.attr = value
+            return item
+
+        def parse(self, value):
+            return value.attr
+
+    def setUp(self):
+        class Cls(xso.XSO):
+            TAG = ("uri:foo", "p")
+
+            values = xso_model.ChildValueList(self.ChildValueType)
+
+        self.Cls = Cls
+
+    def test_is_prop_base(self):
+        self.assertTrue(issubclass(
+            xso_model.ChildValueList,
+            xso_model._ChildPropBase
+        ))
+
+    def test_get_tag_map_and_classes(self):
+        self.assertDictEqual(
+            {
+                self.ChildXSO.TAG: self.ChildXSO
+            },
+            self.Cls.values.get_tag_map()
+        )
+        self.assertSetEqual(
+            self.Cls.values._classes,
+            {self.ChildXSO}
+        )
+
+    def test_registered_at_class_child_map(self):
+        self.assertIn(
+            self.ChildXSO.TAG,
+            self.Cls.CHILD_MAP
+        )
+        self.assertIs(
+            self.Cls.CHILD_MAP[self.ChildXSO.TAG],
+            self.Cls.values
+        )
+        self.assertIn(
+            self.Cls.values,
+            self.Cls.CHILD_PROPS
+        )
+
+    def test_get_on_class_returns_descriptor(self):
+        desc = xso_model.ChildValueList(self.ChildValueType)
+
+        class Cls(xso.XSO):
+            values = desc
+
+        self.assertIs(desc, Cls.values)
+
+    def test_initial_value_is_a_mutable_sequence(self):
+        obj = self.Cls()
+        self.assertIsInstance(
+            obj.values,
+            collections.abc.MutableSequence)
+        self.assertSequenceEqual(obj.values, [])
+
+    def test_get_returns_per_instance_list(self):
+        obj1 = self.Cls()
+        obj2 = self.Cls()
+        self.assertIs(obj1.values, obj1.values)
+        self.assertIsNot(obj1.values, obj2.values)
+        self.assertIs(obj2.values, obj2.values)
+
+    def test_set_is_prohibited(self):
+        obj = self.Cls()
+        with self.assertRaises(AttributeError):
+            obj.values = [1]
+
+    def test_to_sax_calls_format(self):
+        obj = self.Cls()
+        obj.values.append(10)
+
+        base = unittest.mock.Mock()
+
+        with contextlib.ExitStack() as stack:
+            stack.enter_context(unittest.mock.patch.object(
+                self.ChildValueType,
+                "format",
+                new=base.format
+            ))
+
+            self.Cls.values.to_sax(obj, base.dest)
+
+        self.assertSequenceEqual(
+            base.mock_calls,
+            [
+                unittest.mock.call.format(10),
+                unittest.mock.call.format().unparse_to_sax(base.dest)
+            ]
+        )
+
+    def test_from_events_uses__process_and_parse(self):
+        obj = self.Cls()
+
+        def process(mock, *args, **kwargs):
+            return mock(*args, **kwargs)
+            yield None
+
+        base = unittest.mock.Mock()
+        process_mock = base.process
+        base.process = functools.partial(process, process_mock)
+
+        with contextlib.ExitStack() as stack:
+            stack.enter_context(unittest.mock.patch.object(
+                self.ChildValueType,
+                "parse",
+                new=base.parse
+            ))
+
+            stack.enter_context(unittest.mock.patch.object(
+                self.Cls.values,
+                "_process",
+                new=base.process
+            ))
+
+            with self.assertRaises(StopIteration):
+                gen = self.Cls.values.from_events(
+                    obj,
+                    base.ev_args,
+                    base.ctx,
+                )
+                next(gen)
+
+            with self.assertRaises(StopIteration):
+                gen = self.Cls.values.from_events(
+                    obj,
+                    base.ev_args,
+                    base.ctx,
+                )
+                next(gen)
+
+        calls = list(base.mock_calls)
+        self.assertSequenceEqual(
+            calls,
+            [
+                unittest.mock.call.process(obj, base.ev_args, base.ctx),
+                unittest.mock.call.parse(process_mock()),
+                unittest.mock.call.process(obj, base.ev_args, base.ctx),
+                unittest.mock.call.parse(process_mock()),
+            ]
+        )
+
+        self.assertSequenceEqual(
+            obj.values,
+            [
+                base.parse(),
+                base.parse(),
+            ]
+        )
+
+    def tearDown(self):
+        del self.Cls
+
+
+class TestChildValueMap(unittest.TestCase):
+    class ChildXSO(xso.XSO):
+        TAG = ("uri:foo", "foo")
+
+        key = xso.Attr("k", type_=xso.Integer())
+        value = xso.Attr("v", type_=xso.Integer())
+
+    class ChildValueType(xso.AbstractType):
+        @classmethod
+        def get_formatted_type(cls):
+            return TestChildValueMap.ChildXSO
+
+        def format(self, item):
+            key, value = item
+            item = TestChildValueMap.ChildXSO()
+            item.key = key
+            item.value = value
+            return item
+
+        def parse(self, item):
+            return item.key, item.value
+
+    def setUp(self):
+        class Cls(xso.XSO):
+            TAG = ("uri:foo", "p")
+
+            values = xso_model.ChildValueMap(self.ChildValueType)
+
+        self.Cls = Cls
+
+    def test_is_child_prop_base(self):
+        self.assertTrue(issubclass(
+            xso_model.ChildValueMap,
+            xso_model._ChildPropBase
+        ))
+
+    def test_get_tag_map_and_classes(self):
+        self.assertDictEqual(
+            self.Cls.values.get_tag_map(),
+            {
+                self.ChildXSO.TAG: self.ChildXSO,
+            }
+        )
+        self.assertSetEqual(
+            self.Cls.values._classes,
+            {self.ChildXSO}
+        )
+
+    def test_registered_at_class_child_map(self):
+        self.assertIn(
+            self.ChildXSO.TAG,
+            self.Cls.CHILD_MAP
+        )
+        self.assertIs(
+            self.Cls.CHILD_MAP[self.ChildXSO.TAG],
+            self.Cls.values
+        )
+        self.assertIn(
+            self.Cls.values,
+            self.Cls.CHILD_PROPS
+        )
+
+    def test_get_on_class_returns_descriptor(self):
+        desc = xso_model.ChildValueList(self.ChildValueType)
+
+        class Cls(xso.XSO):
+            values = desc
+
+        self.assertIs(desc, Cls.values)
+
+    def test_initial_value_is_a_mutable_mapping(self):
+        obj = self.Cls()
+        self.assertIsInstance(
+            obj.values,
+            collections.abc.MutableMapping)
+        self.assertSequenceEqual(obj.values, {})
+
+    def test_get_returns_per_instance_dict(self):
+        obj1 = self.Cls()
+        obj2 = self.Cls()
+        self.assertIs(obj1.values, obj1.values)
+        self.assertIsNot(obj1.values, obj2.values)
+        self.assertIs(obj2.values, obj2.values)
+
+    def test_mapping_type_argument(self):
+        mock = unittest.mock.Mock()
+
+        class Cls(xso.XSO):
+            TAG = ("uri:foo", "p")
+
+            values = xso_model.ChildValueMap(
+                self.ChildValueType,
+                mapping_type=mock
+            )
+
+        obj = Cls()
+        mapping = obj.values
+        self.assertEqual(mapping, mock())
+
+    def test_mapping_type_not_called_if_attr_already_created(self):
+        mock = unittest.mock.Mock()
+
+        class Cls(xso.XSO):
+            TAG = ("uri:foo", "p")
+
+            values = xso_model.ChildValueMap(
+                self.ChildValueType,
+                mapping_type=mock
+            )
+
+        obj = Cls()
+        obj.values
+        obj.values
+
+        self.assertSequenceEqual(
+            mock.mock_calls,
+            [
+                unittest.mock.call(),
+            ]
+        )
+
+    def test_set_is_prohibited(self):
+        obj = self.Cls()
+        with self.assertRaises(AttributeError):
+            obj.values = {}
+
+    def test_to_sax_calls_format(self):
+        obj = self.Cls()
+        obj.values[10] = 20
+
+        base = unittest.mock.Mock()
+
+        with contextlib.ExitStack() as stack:
+            stack.enter_context(unittest.mock.patch.object(
+                self.ChildValueType,
+                "format",
+                new=base.format
+            ))
+
+            self.Cls.values.to_sax(obj, base.dest)
+
+        self.assertSequenceEqual(
+            base.mock_calls,
+            [
+                unittest.mock.call.format((10, 20)),
+                unittest.mock.call.format().unparse_to_sax(base.dest)
+            ]
+        )
+
+    def test_from_events_uses__process_and_parse(self):
+        obj = self.Cls()
+
+        def process(mock, *args, **kwargs):
+            return mock(*args, **kwargs)
+            yield None
+
+        base = unittest.mock.Mock()
+        process_mock = base.process
+        base.process = functools.partial(process, process_mock)
+
+        with contextlib.ExitStack() as stack:
+            stack.enter_context(unittest.mock.patch.object(
+                self.ChildValueType,
+                "parse",
+                new=base.parse
+            ))
+
+            stack.enter_context(unittest.mock.patch.object(
+                self.Cls.values,
+                "_process",
+                new=base.process
+            ))
+
+            base.parse.return_value = (1, "a")
+            with self.assertRaises(StopIteration):
+                gen = self.Cls.values.from_events(
+                    obj,
+                    base.ev_args,
+                    base.ctx,
+                )
+                next(gen)
+
+            base.parse.return_value = (2, "b")
+            with self.assertRaises(StopIteration):
+                gen = self.Cls.values.from_events(
+                    obj,
+                    base.ev_args,
+                    base.ctx,
+                )
+                next(gen)
+
+        calls = list(base.mock_calls)
+        self.assertSequenceEqual(
+            calls,
+            [
+                unittest.mock.call.process(obj, base.ev_args, base.ctx),
+                unittest.mock.call.parse(process_mock()),
+                unittest.mock.call.process(obj, base.ev_args, base.ctx),
+                unittest.mock.call.parse(process_mock()),
+            ]
+        )
+
+        self.assertDictEqual(
+            obj.values,
+            {
+                1: "a",
+                2: "b"
+            }
+        )
+
+    def tearDown(self):
+        del self.Cls
+
+
+class TestChildTextMap(unittest.TestCase):
+    def test_init(self):
+        base = unittest.mock.Mock()
+        with contextlib.ExitStack() as stack:
+            stack.enter_context(unittest.mock.patch.object(
+                xso_model.ChildValueMap,
+                "__init__",
+                new=base.ChildValueMap
+            ))
+
+            stack.enter_context(unittest.mock.patch(
+                "aioxmpp.xso.types.TextChildMap",
+                new=base.TextChildMap
+            ))
+
+            obj = xso_model.ChildTextMap(base.xso)
+
+        calls = list(base.mock_calls)
+        self.assertSequenceEqual(
+            calls,
+            [
+                unittest.mock.call.TextChildMap(base.xso),
+                unittest.mock.call.ChildValueMap(
+                    base.TextChildMap(),
+                    mapping_type=structs.LanguageMap,
+                ),
+            ]
+        )
+
+        self.assertIsInstance(
+            obj,
+            xso_model.ChildTextMap
+        )
+
+
 class TestXSOParser(XMLTestCase):
     def run_parser(self, classes, tree):
         results = []